move legacy module loading code into a separate file
