+++ lib/ace/ace.js
@@ -36,8 +36,7 @@
 
 define(function(require, exports, module) {
 "use strict";
-
-require("./lib/fixoldbrowsers");
+"include loader_build";
 
 var dom = require("./lib/dom");
 var event = require("./lib/event");
@@ -59,15 +58,6 @@ require("./ext/error_marker");
 
 exports.config = require("./config");
 
-/**
- * Provides access to require in packed noconflict mode
- * @param {String} moduleName
- * @returns {Object}
- **/
-exports.require = require;
-
-if (typeof define === "function")
-    exports.define = define;
 
 /**
  * Embeds the Ace editor into the DOM, at the element provided by `el`.
