+++ lib/ace/config.js
@@ -39,10 +39,6 @@ var AppConfig = require("./lib/app_config").AppConfig;
 
 module.exports = exports = new AppConfig();
 
-var global = (function() {
-    return this || typeof window != "undefined" && window;
-})();
-
 var options = {
     packaged: false,
     workerPath: null,
@@ -113,12 +109,12 @@ exports.setModuleUrl = function(name, subst) {
 var loader = function(moduleName, cb) {
     if (moduleName == "ace/theme/textmate")
         return cb(null, require("./theme/textmate"));
-    return console.error("loader is not configured")
-}
+    return console.error("loader is not configured");
+};
 
 exports.setLoader = function(cb) {
-    loader = cb
-}
+    loader = cb;
+};
 
 exports.$loading = {};
 exports.loadModule = function(moduleName, onLoad) {
@@ -175,63 +171,6 @@ var reportErrorIfPathIsNotConfigured = function() {
         reportErrorIfPathIsNotConfigured = function() {};
     }
 };
-
-// initialization
-function init(packaged) {
-    if (!global || !global.document)
-        return;
-    
-    options.packaged = packaged || require.packaged || module.packaged || (global.define && define.packaged);
-
-    var scriptOptions = {};
-    var scriptUrl = "";
-
-    // Use currentScript.ownerDocument in case this file was loaded from imported document. (HTML Imports)
-    var currentScript = (document.currentScript || document._currentScript ); // native or polyfill
-    var currentDocument = currentScript && currentScript.ownerDocument || document;
-    
-    var scripts = currentDocument.getElementsByTagName("script");
-    for (var i=0; i<scripts.length; i++) {
-        var script = scripts[i];
-
-        var src = script.src || script.getAttribute("src");
-        if (!src)
-            continue;
-
-        var attributes = script.attributes;
-        for (var j=0, l=attributes.length; j < l; j++) {
-            var attr = attributes[j];
-            if (attr.name.indexOf("data-ace-") === 0) {
-                scriptOptions[deHyphenate(attr.name.replace(/^data-ace-/, ""))] = attr.value;
-            }
-        }
-
-        var m = src.match(/^(.*)\/ace(\-\w+)?\.js(\?|$)/);
-        if (m)
-            scriptUrl = m[1];
-    }
-
-    if (scriptUrl) {
-        scriptOptions.base = scriptOptions.base || scriptUrl;
-        scriptOptions.packaged = true;
-    }
-
-    scriptOptions.basePath = scriptOptions.base;
-    scriptOptions.workerPath = scriptOptions.workerPath || scriptOptions.base;
-    scriptOptions.modePath = scriptOptions.modePath || scriptOptions.base;
-    scriptOptions.themePath = scriptOptions.themePath || scriptOptions.base;
-    delete scriptOptions.base;
-
-    for (var key in scriptOptions)
-        if (typeof scriptOptions[key] !== "undefined")
-            exports.set(key, scriptOptions[key]);
-}
-
-exports.init = init;
-
-function deHyphenate(str) {
-    return str.replace(/-(.)/g, function(m, m1) { return m1.toUpperCase(); });
-}
 
 exports.version = "1.7.1";
